pyproject.toml
[tool.poetry]
name = "vlm-creation"
version = "0.1.0"
description = ""
authors = ["Anton Shiryaev <med.phisiker@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.5.1"
transformers = "^4.48.1"
accelerate = "^1.3.0"
huggingface-hub = "^0.27.1"
peft = "^0.14.0"
datasets = "^3.2.0"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
